#45 - updated Feature Collections
